Sports Stats Project/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
nba_api
anthropic
python-dotenv